requirements.txt
blitzdb
aiohttp
pymongo
requests
py-cpuinfo